add theme3

+++ copy-replace.js
@@ -2,6 +2,11 @@ const fs = require("fs-extra");
 require("dotenv").config();
 const destPath = "./public";
 try {
+  if (process.env.REACT_APP_VERSION_CODE === undefined) {
+    throw new Error(
+      "REACT_APP_VERSION_CODE is not defined, Please check .env file!\n"
+    );
+  }
   fs.removeSync(destPath);
   console.log("Directory successfully removed.");
 
@@ -10,5 +15,9 @@ try {
   fs.copySync(srcPath, destPath);
   console.log("Directory successfully copied.");
 } catch (err) {
-  console.error("Error", err);
+  console.error(
+    "Error " +
+      "REACT_APP_VERSION_CODE is not defined, Please check .env file!\n",
+    err
+  );
 }
